improve: Borrow classes from pglast and implement a simple visitor
pglast has 4 classes that represent a query dag. These do not have a
function to accept a visitor. It was not possible to inherit as a
factory method and a few other functions have hard-coded the class
names. So a copy of the classes with an accepting function has been
added.

A simple visitor that prints the node has been implemented as POC.